feat(tools-menu): created a top menu that includes multible tools
- Changed the UI of the editor
- Changed the file structure
- Created different components for different types of editor tools
- Minimized the amount of tools in the bubble menu
