+++ src/styles/components/DropDown.css
@@ -1,5 +1,26 @@
 .noqta-drop-menu {
 	position: relative;
+	font-size: 0.8rem;
+}
+
+.noqta-drop-menu-header {
+	display: flex;
+	align-items: center;
+	background: var(--noqta-theme-background-primary);
+	border: 2px solid var(--noqta-theme-border-primary);
+	border-radius: 0.5rem;
+	cursor: pointer;
+	user-select: none;
+	padding: 0.5rem;
+	gap: 0.25rem;
+	text-wrap: nowrap;
+	transition: all 0.2s ease-in-out;
+}
+
+.noqta-drop-menu-active-option {
+	width: 5rem;
+	text-overflow: ellipsis;
+	overflow: hidden;
 }
 
 .noqta-drop-menu-content {
@@ -24,10 +45,10 @@
 	padding: 0.5rem;
 	cursor: pointer;
 	border-radius: 0.25rem;
-	font-size: 0.8rem;
 	transition: all 0.2s ease-in-out;
 }
 
+.noqta-drop-menu-header:hover,
 .noqta-drop-menu-option:hover {
 	background: var(--noqta-theme-background-hover, --noqta-theme-background-primary);
 }
