+++ src/styles/components/NoqtaEditor.css
@@ -1,5 +1,6 @@
-#noqta-editor {
+#editor-container {
 	background-color: var(--noqta-theme-background-primary);
 	color: var(--noqta-theme-text-primary);
 	border-color: var(--noqta-theme-border-primary, transparent);
+	border-radius: 0.5rem;
 }
